requirements.txt
fastapi
uvicorn
boto3
aerospike
python-dotenv
requests
langchain
langchain-aws
langgraph
auth0-python
